qdoc: Introduce NativeEnum container
The \qmlenumeratorsfrom command is used for replicating native (C++)
enum documentation on QML type reference. When parsing the command,
QDoc searches for an EnumNode to extract the documentation from.

Previously, this command worked only in a \qmlproperty topic, and
consequently, the native enum information was also stored in a
QmlPropertyNode.

With the introduction of the \qmlenum topic however, this same
mechanism must work in two different Node subclasses, QmlEnumNode and
QmlPropertyNode.

Introduce a new NativeEnum container that stores an EnumNode pointer
and an optional enumerator prefix, and add it as a member of the
two node types. The container also provides a resolve() function to
search for the specified enum node.

To access the new container, add a NativeEnumInterface class,
implemented by the Nodes that can be associated with a native enum.

Rename Generator::generateEnumValuesForQmlProperty(), as it now
works with any node that implements the new interface.

Task-number: QTBUG-134806
Change-Id: I5e043edc59364b23640af15ed0019eaea23b3624
Reviewed-by: Paul Wicking <paul.wicking@qt.io>