lupdate: add used namespaces of the current file
The namespaces used in a parent scope are already propagated to the
child namespace. However, if the current namespace is first defined
in a different file, it may have a different parent than the one
in the current file.

This patch adds the used namespaces in the current file when entering
a namespace that is defined in another file and has another parent.

Fixes: QTBUG-136768
Pick-to: 6.9
Change-Id: Ide52eb1ce6850ce38d543ebe6837019ebd1790d1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>