linguist: add new example for Qt i18n in plurals and time formats
This patch adds an example to demonstrate use of plural
forms, and showcasing the localized time formats.
The example is a simple clock in Qml that counts the
seconds (plural forms) and presents the time and date
(time formats), in the localized format.
It also show cases the best practices in integrating Qt
i18n in CMake.

Fixes: QTBUG-71856
Change-Id: If28950a1d3e6be4ff2a49d28b13ba09eac472281
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 99e98ea257778f0144eb9413de425b39c61762ab)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>