linguist: add new example for Qt i18n in runtime language changes
This patch adds an example to demonstrate switching languages in
applications at runtime.
The example is a simple clock in Qt Quick that counts the
seconds (plural forms) and presents the time and date.
The application allows the user to switch the language
of the UI at runtime by clicking on a button and choosing
a new language from the menu.
It also show cases the best practices in integrating Qt
i18n in CMake.

Fixes: QTBUG-71856
Pick-to: 6.10
Change-Id: I8ba866902e690b52a41cafbbe5e1ce88b3d63ebd
Reviewed-by: Kai Köhne <kai.koehne@qt.io>