linguist: add new example for Qt i18n for ID based translation
This patch adds an example to demonstrate ID based translation
in Qt. The example is a simple clock in Qml that counts the
seconds (plural forms) and presents the time and date.
The application allows the user to change the time zone of
the clock, by opening a dialog (in C++ widgets) that contains
a list of time zones. The example shows how the same translation
can be reused with the same ID across Qml and Widgets.
It also show cases the best practices in integrating Qt
i18n for ID based translation in CMake.

Pick-to: 6.10
Fixes: QTBUG-71856
Change-Id: I40b707afc3fa1e7e57c35507b0a73c6e48ca1139
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>