QDoc: Don't auto-link to deprecated functions
Function auto-links that resolve to deprecated functions prevent valid
alternatives from being found. By discarding deprecated functions we at
least make it possible that a suitable target can be found.

Auto-links to deprecated functions are not linked in any case, so we do
not lose anything by doing this. It is possible that unrelated,
identically-named functions might be found and linked.

Fixes: QTBUG-137147
Change-Id: I500c9bf750a54a1e9807b92eb8f3391a2ad42a3a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>