QDoc: Promote documentation in headers feature
The `documentationinheaders` feature was introduced to QDoc as a
technical preview feature with Qt 6.9. Some minor issues have been
reported and addressed since its release. Given the current state,
we (the development team) feel the feature is mature enough to warrant
being considered a stable QDoc feature. This change "promotes" the
feature from technical preview to regular feature.

Specifically, the change removes the "technical preview" phrase from the
documentation of the feature. As it was missing, a mention of the
version in which the feature was introduced is added as well.

Change-Id: Id2eb1baea08bdf7632f04608b685e98ba5ac3609
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit f7efd5d856656df7293629196b42fbef18ef9efd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>