QDoc: Avoid accidental nullptr dereferencing
Coverity complains that `parent` can be dereferenced in the
QDocIndexFiles::readIndexSection function, and indeed it is possible to
create a situation where that happens. This change ensures that the
nullptr check applies equally to the various 'or' conditions in the if
expression.

Coverity-Id: 480819
Pick-to: 6.9 6.8
Change-Id: I8c6a9095bc837c772b9a9d699bb5dbe400eb56de
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 54423aeadffced7db537a469474bc13f6a855234)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>