Buddy: search for a buddy bottom-up rather than top-down
As it stood, we would search for a label's buddy from the top level
widget and down, based on the object name. The problem with that approach
is that since the same form can be instantiated onto multiple widgets
inside the same window, there can exist multiple buddies with the same
name as well. As a result, the form builder would end up resolving the
same buddy instance for the all the form instances.

Rather than searching for the buddy top-down, this patch will instead
search for it bottom-up. We assume then, that the buddy closest to
the label is the correct one to use.

Fixes: QTBUG-96693
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Iae7c8865563304810ed5f905c6ad45d7b796b7a5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>