QDoc: Warn about self-linking in \sa commands
If a \sa command contains a link to the document hosting it, such as
documentation for a method containing a link to the method, it was
the case that QDoc would simply show the text without self-linking or
warning.

This change performs checks on each link to determine if self-linking
is occurring and warns if so.

Fixes: QTBUG-137048
Change-Id: I358f2567d3748375aff3daf68ec4bfdd48a5b0a3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit c884af4b3ad337bb90b54d10df9902befaa8f25b)
Reviewed-by: Topi Reinio <topi.reinio@qt.io>