CMake: Use qtbase autogen build dir query function for lupdate
Because the path to the dir might change when using a different cmake
autogen intermediate dir strategy, we should query the value using the
qtbase function, rather than expecting it to be the default binary
dir.

We keep the genex approach for when the target property might be set,
because it might be set after the call to qt6_add_lupdate.

Amends d39e9f21c98eb4788f3434f139b828862b2d8823

Pick-to: 6.8 6.9 6.10
Task-number: QTBUG-110696
Change-Id: Ic2e4a210b95e23fa21563f7995ad9b1dfd3c3e6c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>