QDoc: Fall back to deprecated functions when auto-linking
Previously, the auto-linking strategy would find deprecated functions
that shadowed functions with the same name - see QTBUG-137147.
The fix for this prevented linking to deprecated functions, but this
caused warnings when the deprecated functions were the intended link
targets.

This change enables linking to deprecated functions as a last resort,
when the only suitable function is deprecated.

Fixes: QTBUG-138557
Pick-to: 6.10
Change-Id: Ie20a282d0a897d987f66a5b550e9d54923d65e7d
Reviewed-by: Kai Köhne <kai.koehne@qt.io>