QDoc: Add an option to log warnings to a file
For large documentation sets or in CI environments (e.g. COIN), warnings
can be numerous and scroll by too quickly to be useful. This makes it
cumbersome to systematically track and fix documentation issues.

This commit introduces a new configuration variable, `logwarnings`.
When set to `true` in a .qdocconf file, QDoc will write all warning
messages to a log file named `<project>-qdoc-warnings.log` in the
specified output directory. The warnings are also written to
stderr (existing behavior). As messages to stderr are flushed
immediately, the implementation doesn't continuously flush messages to
the warning log file.

To provide context for reproducibility, the command-line arguments
used to invoke QDoc are logged at the top of the file.

[ChangeLog][QDoc] QDoc can now write warnings to a file. To enable this
feature, set the configuration variable `logwarnings` to `true`.

Fixes: QTBUG-138811
Change-Id: If9d5d3ccd0d75265b317ce17370e99b2d418ed02
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>