linguist doc: update the documentation to match with the new icons
Linguist has been updated with new icons and marks for translations.
The new patch reflects the new marks in the documentation as well. Also
the screenshots are updated to reflect the new look.
In addition, the new label feature is included in the screenshots
and is being documented.

Fixes: QTBUG-138631
Change-Id: Ibe2ef64d63c531fdfb177a918ed25bb2e4f8326e
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit e779c62c560962926c6c1e52d7ed65ffcdf6c74c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>