QDoc: Add test coverage support using gcov/lcov
This change introduces a mechanism to generate test coverage reports
for QDoc using gcov/lcov. This will aid QDoc's developers track test
quality and identify untested areas of the codebase.

The feature requires the 'lcov' and 'genhtml' tools to be installed.
The feature is enabled for developer builds where QtBase is configured
with the 'coverage_gcov' feature and when QtTools is configured to build
tests. It's developed for internal use and on Linux only; MSVC doesn't
work. On macOS, there be dragons -- you have been warned.

Three new build targets are added:
- qdoc_test_with_coverage: A convenience target that builds, runs
  tests, collects data, and generates the final HTML report. This is
  the primary target for developers.
- qdoc_coverage_collect: Collects coverage data from the current test
  run.
- qdoc_coverage_report: Generates the HTML report from collected data.

Usage:
1. Configure the QtBase build with tests and coverage enabled:
   configure -developer-build -coverage gcov -make tests
2. Configure QtTools.
3. Build the new target:
   cmake --build . --target qdoc_test_with_coverage
4. Open the report:
   <build-dir>/coverage_report/index.html

Fixes: QTBUG-124895
Change-Id: I59957fb69a7ae6db2a57345d41eb30de7cb115c9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>