QDoc: Add \overload handling to DocBook generator
This change adds generation of overload notifications in DocBook output,
matching functionality already present in other output formats.
When a function has overloads and contains an overload command,
display appropriate text with links to the base function.

The implementation adds a case to ``generateAddendum()` that generates
either "This is an overloaded function" or "This function overloads
[target]" messages depending on the context. The code includes automatic
link resolution for overload targets and qualifies function names with
parent class context when needed to improve link accuracy.

Fixes: QTBUG-138890
Change-Id: Idef429e707583cce2195f83ad7b3007ad3177a30
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>