linguist: check the size of the group tab before switching to it
When navigating on Qt Linguist to a previous message from index 0,
linguist automatically switches to the inactive tab and continues
the navigation from there. If the inactive group tab is empty
(label or context), then a non-valid index is reached.
This patch checks whether the inactive tab is empty before switching
to it.

Fixes: QTBUG-138910
Pick-to: 6.10
Change-Id: Ibffb41e8fe0a756014513a06436d9ffde2c0748c
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>