QDoc: Add `\overload primary` to set primary overload
When `\overload` is used with an unqualified function name, QDoc uses
an automatic sorting algorithm to determine the primary function to link
to. In cases with multiple similar overloads, this algorithm can fail
to select a single, unambiguous target, resulting in broken links. The
only workaround is to use a verbose, fully qualified function signature.

This change introduces a `primary` argument to the `\overload` command,
allowing authors to explicitly designate a function as the primary one
in an overload set. The sorting logic is updated to prioritize this
designated function, resolving ambiguity and ensuring unqualified
`\overload` commands link correctly.

To prevent incorrect usage, a warning is now emitted if multiple
functions in the same overload set are marked as primary.

[ChangeLog][QDoc] You can now specify what the primary signature is for
a set of overloads with `\overload primary`. Refer to the QDoc manual
for detailed examples and documentation.

Fixes: QTBUG-138586
Fixes: QTBUG-138955
Change-Id: I351b55ac9185d2b74da4da731362757592f244e0
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>