qdoc: Add an optional 'offset' argument to Atom::find()
When searching the linked list of Atoms for a specific atom type, a
useful piece of information in certain cases is the offset between
the current and matched atom. One use case is skipping all atoms
between specified [left, right] atom types.

Implement this feature in Atom::find() to avoid the call sites having
to calculate the offset.

Change-Id: Ia6c8408117dcdd9f92c9cd8620d96bb65f397fcd
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: David Boddie <david.boddie@qt.io>