QDoc: Introduce \toc .. \endtoc, \tocentry commands
Introduce \toc and \endtoc commands that can be used for specifying a
list of sub-topics for a topic (page). Each entry is specified using
a \tocentry command. Sub-topics can specify their own \toc,
resulting in a nested hierarchy of table of contents entries
for the documentation project.

QDoc parses the contents of the command block separately using a
new class `TOCWriter` in generate phase. To generate a TOC hierarchy,
a root topic (index page) is required. This is read from
`navigation.landingpage`, or as a fallback, `navigation.homepage`.
If no such configuration exists or the root topic has no \toc,
TOCWriter does nothing.

To prevent infinite loops due to circular toc entries, TOCWriter
imposes a maximum nesting level of 16 for the hierarchy.

As output, TOCWriter generates a <project>_toc.xml file. The syntax
is similar to .qhp files' <toc> element, but with a per-topic control
of the hierarchy, maintained in documentation sources.

For a modular build such as the Qt framework, the final output is
a set of XML files, each defining the module's table of contents.
These files can then be used for constructing the hierarchy for the
top-level navigation sidebar.

[ChangeLog][QDoc] Added support for generating a table of contents
in XML format for the documentation project using \toc commands.

Fixes: QTBUG-138684
Change-Id: Ib7e7466def4521c4827a675080584fcd66c5e0da
Reviewed-by: David Boddie <david.boddie@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>