QDoc: Add \qmlsingletontype command
QDoc supports `\qmltype` for documenting QML types. However, it does not
let authors explicitly mark a type as a singleton in documentation.

This change adds a new `\qmlsingletontype` command. It works the same as
`\qmltype`, but always marks the type as a singleton. The HTML and
DocBook generators now display "(Singleton)" in the title and include a
note explaining the singleton behavior. The index files also preserve a
singleton attribute. Tests and documentation are updated to cover the
new command.

[ChangeLog][QDoc] QDoc now has the \qmlsingletontype command, which you
can use to document a QML type as a QML singleton. Refer to the QDoc
manual for further details.

Task-number: QTBUG-51584
Change-Id: I4effda22f031cbcff7a871fbe8f460b0a3f78f3f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>