QDoc: Only assign overload numbers to documented functions
Functions were previously assigned numbers whether they were documented
or not. This led to links containing overload numbers that did not
reflect the number of overloads in the documentation. In the minimal
case, a single function with no visible overloads would be assigned a
number.

The extraction mark is also affected in the minimal case since the
[overload1] text is included for primary overloads.

Task-number: QTBUG-139262
Pick-to: 6.10
Change-Id: If4678721c4899525ee9971db48f4c41c9513f464
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>