QDoc: Change title of How to Resolve QDoc Warnings
The doc team is trying to implement a version of the diataxis framewrork
to doc.qt.io.

The QDoc manual is being updated with this framework in mind.

This patch changes the title of the 'How to Resolve QDoc Warnings'
page to 'Troubleshooting QDoc Warnings' because the 'How-to' title
should only be used in a specific context according to the framework.

This patch also updates the links to this page.

Fixes: QTBUG-139449
Pick-to: 6.10
Change-Id: Ibb6e2a6eddb1a0434b8b7280e1b37e048250c07a
Reviewed-by: Topi Reinio <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>