CMake: Drop 'git' in VERSION argument for dev builds
LLVM has a strict compatibility check in the CMake packages they
provide. Therefore, we cannot pass a non-conformant version string,
such as `21.0.0git`, to the `find_package()` call for Clang or LLVM.

QDoc carries a fork of `clang/AST/QualTypeNames` due to upstream changes
between LLVM 15 and LLVM 16 that were deemed incompatible with QDoc's
needs. As this class sees upstream modifications that break compilation,
there's a need to bisect LLVM to track down the offending upstream
change, so that QDoc's implementation can be modified to work correctly
alongside newer versions of LLVM. These builds contain version strings
that contain 'git' as a suffix, and thus break the developer experience
in qttools.git.

To mitigate this problem, this change strips the suffix by a simple
regular expression based string replacement. The new behavior is gated
with `QT_FEATURE_developer_build` (the feature flag defined by a
`-developer-build` configuration of qtbase.git). The rationale is that,
as this is strictly speaking a convenience for me, myself, and I, it
should not be the default behavior of the CMake script, and if run in
an environment that's a developer build, it's reasonable to expect that
the developer responsible for configuring qttools.git is able to find
and understand why this happens.

This change extends the logic implemented in
914e00bbd6b35a376b6f71c031d9469bd4d9f672 and is, as such, cherry-picked
as far back as to the branches that change targeted.

Task-number: QTBUG-139407
Pick-to: 6.8
Change-Id: Id5dea0678037cc5f86343b7aa0421b27c6ad13c0
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: David Boddie <david.boddie@qt.io>
(cherry picked from commit 39f0f4863d9a5d5143ed8b04a150865196084386)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit f61c47ef046de92420ce86902e4ea781b11fef76)
Reviewed-by: Paul Wicking <paul.wicking@qt.io>