QDoc: Store declared type in index file
In certain situations, authors may prefer to use `auto` instead of the
actual type of a function as argument QDoc's `\fn`-command. This makes
QDoc output `auto` instead of the full type, which can be used to hide
certain implementation details. QDoc stores this as `declaredReturnType`
on function nodes.

However, the declared type is not written to index files. This means
that modules that read the index file as a dependency always get the
de-sugared type, despite the author's intentions.

This change extends QDoc's index file by adding a new attribute,
`declaredtype`. If present, the value is now written to and read from
.index-files by QDoc.

The change is cherry-picked back to 6.8, where the declared return type
override feature was added.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-139614
Change-Id: Ie73c3db070943aa91e63d92d3e33e72d9579a7ee
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>