QDoc: Honor includeprivate* settings in summary links
QDoc suppresses links for private nodes even when they're included in
the documentation. The result is that the entries in summary sections
for private members on C++ reference pages don't contain links to the
detailed description.

This change adds `XmlGenerator::shouldIncludePrivateNode()` as a helper
for checking the configuration of the `includeprivate*` settings, and
results in QDoc correctly generating links for private members based on
the configuration. The helper is called from
`XmlGenerator::linkForNode()`.

It also updates the generated output tests to reflect the new behavior.

Task-number: QTBUG-58805
Task-number: QTBUG-63977
Task-number: QTBUG-99553
Task-number: QTBUG-114877
Change-Id: I003f3deae8ee85fe2c6cc8bcde9c9fb157ac9066
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>