QDoc: Test hidden friends in output
Recent QDoc changes enabled documentation for private members. As a
result, hidden friends (friend functions declared in a class’s private
section but visible via ADL) are generated under “Related Non-Members”.

This change adds a dedicated test case to demonstrate that behavior
when `includeprivate` is not set, and extends the private members test
to show the same functions also appear when `includeprivate` is set.

Fixes: QTBUG-99553
Change-Id: I571fea315fcda560aace6d523cad413a5929261f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>