QDoc: Don’t force private access for \internal nodes
When processing documentation, QDoc flags nodes documented as
`\internal` as internal. However, the procedure for setting the flag on
the node also sets the access modifier to `private`, which indicates
that it belongs to a `private` section in a C++ class.

Now that QDoc can generate documentation for private members, coupling
the "internal" state to the access modifier is incorrect, as internal
items could end up being rendered under the private sections when shown.

This change modifies Node::markInternal and
Aggregate::markUndocumentedChildrenInternal to only set status as
internal.

Task-number: QTBUG-58805
Task-number: QTBUG-63977
Task-number: QTBUG-99553
Task-number: QTBUG-114877
Task-number: QTBUG-139658
Change-Id: Icd41267e350f88677fc6d9defa44aa04c0836e90
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>