QDoc: Hide internal nodes from the search for child nodes in aggregates
Types like qsizetype, QVariantList, QVariantHash and QVariantMap were
not always linked to their documentation when shown in function
signatures. The cause for qsizetype was other identically-named internal
nodes being chosen before it during a search for child nodes in
aggregates.

By excluding internal nodes from the search, a number of types are now
found and linked.

Fixes: QTBUG-140110
Pick-to: 6.10
Change-Id: If37f343a4e4abde90e3047e7ff44811580f3752b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>