QDoc: Report unresolved \relates targets
QDoc silently created 'Proxy' nodes for entities that use \relates
command to associate themselves with another aggregate (class,
namespace, or header file), if the aggregate specified in \relates
argument could not be found.

While this behavior is meant to allow documenting e.g. members of
a single namespace across multiple modules, it also means that
a typo in the \relates argument results in a unintended creation of
a proxy page.

Improve the logic by searching trough all trees in the forest for the
\relates aggregate. If one is not found, report it. If an
aggregate is found but comes from an index (read in as a dependency),
proceed to construct a Proxy aggregate as before. Therefore, this
change does not alter the output generated by QDoc, it only adds
reporting.

Add a simple test for an unresolved \relates resulting in a proxy page.

[ChangeLog][QDoc] QDoc now outputs a report if the \relates command
refers to an unknown target.

Fixes: QTBUG-105172
Change-Id: If453908f84b5a2dc887475789ca88337dc4d8d13
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>