QDoc: Test inclusion strategy
A recent series of changes to QDoc introduced a basic inclusion
strategy. Those changes introduced a number of small POD structures. The
behavior was untested.

This change adds unit test coverage for the new PODs, which will allow
catching future regression to the new policy system as it is extended.

Task-number: QTBUG-140077
Change-Id: Id21c851dc6bf0bd389d36113f52667f4e02aa6ee
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>