QDoc: Extract warning log file header function
QDoc can generate a log file that contains all the warnings for a given
run. This log file starts with a header that is the list of arguments
passed to QDoc on the command line.

This change extracts the warning log creation into a separate function,
as a preporatory step for upcoming changes to the warning log header.

Task-number: QTBUG-138846
Change-Id: I8260a742c7ce03d9ae08d78e7bfdc37d8965d7f5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>