QDoc: Avoid generating invalid HTML for breadcrumbs
The breadcrumb navigation for a generated HTML page consist of
a list with list item (<li>) elements.

The opening and closing list elements for breadcrumbs were expected to
be provided by the configuration variables HTML.postheader and
HTML.postpostheader.

If these were not provided (as in our tests), QDoc produced invalid
HTML as <li> elements must be children of <ol>, <ul>, or <menu>.

Provide sensible default values for both configuration variables.

Adjust documentation of HTML.postheader and document
HTML.postpostheader.

For consistency, move the default value for HTML.endheader to
HtmlGenerator::initializeGenerator().

In passing, swap instances of QLatin1String() for ""_L1
StringLiterals in HtmlGenerator::initializeGenerator().

Fixes: QTBUG-140399
Pick-to: 6.10
Change-Id: I85a34160f6715e7a2b95d1f11c62ac15d5b36f40
Reviewed-by: Paul Wicking <paul.wicking@qt.io>