QDoc: Stop setting title for QML types
For obscure historical reasons, QDoc called `setTitle()` in QmlTypeNode
constructor. Setting a title is superfluous as it was always identical
to the Node's name(), i.e, the argument passed to \qmltype command.

PageNode::setTitle() added the node to the parent aggregate's map of
'non-functions' to facilitate searching by title, which made little
sense for type nodes; it only served to cause duplication in the map
because the node was already added there by Aggregate::addChild().

Drop the `title`, `fulltitle`, and `subtitle` index file attributes for
QML type nodes, and have generators use name() in place of fullTitle().

Task-number: QTBUG-140729
Change-Id: I161e097af9208023c427cb531761440da269f2b7
Reviewed-by: Paul Wicking <paul.wicking@qt.io>