QDoc: Override status reporting for function nodes to be consistent
There are two methods that report whether a node is deprecated: status()
and isDeprecated(). Function nodes override isDeprecated() to enable
deprecated properties to override the status of access functions, but
this does not affect the value returned by status(), which still returns
the underlying status of the functions.

This change unifies the two methods for FunctionNode.

Change-Id: I010e68ac3ed75d883c9b1f23d389cab8db701bae
Fixes: QTBUG-140539
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>