QDoc: Add an optional language argument to the code command
The optional argument enables a recognized programming language to be
specified for code marking instead of relying on QDoc's existing rules
for choosing a marker based on context. Alternatively, "text" can be
specified to avoid using any code marking.

Warn about unrecognized languages.

Add to the section in the manual about highlighting in inline code.

[ChangeLog][QDoc] The \code command now accepts an optional argument to
specify the programming language for the code block.

Pick-to: 6.11
Task-number: QTBUG-98981
Change-Id: Id4105ff28fc6a51cafe9f0ba0703139306d7db8c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>