lupdate: Fix spurious warnings with namespaced return types
In the case of '#include<...>' directive, lupdate looks only in
the configured (via -I flag) and known places. Whereas,
for '#include "..."' directives, lupdate look in the
relative / absolute location of the included file.
Because non configured and unknown angle bracket includes
are not processed, their namespaces and types cannot be
qualified by lupdate.
Due to this, if the returning type of a function (that contains tr
calls) is defined in a file included via '#include<...>' directives
instead of '#include "..."' directive, lupdate incorrectly warns
about unknown type (for the returning type).

Since the returning type of a function does not define the context,
it is irrelevant for lupdate to warn about in the first place.
Hence, this patch clears prospectiveContext (that lupdate will try
to qualify later) when the :: chain breaks at a
non-function/non-member-initialization point (i.e., when the next
token is NOT '(', '[', or '='). Such cases do not define the
context of translation functions and are irrelevant for lupdate
to look into.

Fixes: QTBUG-140636
Pick-to: 6.10
Change-Id: I5c2d75ff6ce998a7247cb4d17f80903af9624a27
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>