QDoc: Introduce imagesoutputdir configuration variable
In multi-module documentation builds such as Qt framework docs, it's
not uncommon for two documentation projects to include an image file
whose names clash with each other.

This is not a problem in the default (offline) configuration where
each documentation project uses a dedicated output directory, but
the online configuration uses a shared output directory for all
projects. This led to clashing image files being overwritten.

Introduce a new configuration variable `imagesoutputdir` for setting
a custom output subdirectory for images.

[ChangeLog][QDoc] The output directory for images used in the
documentation can now be configured with 'imagesoutputdir' variable.

Fixes: QTBUG-129446
Change-Id: I2a6e166576bcde4f9a1d9cf1b37f965741b259fa
Reviewed-by: Paul Wicking <paul.wicking@qt.io>