QDoc: Fix --redirect-documentation-to-dev-null
The `--redirect-documentation-to-dev-null` command-line option designed
for the purpose of ensuring that QDoc doesn't generate any files as
output. It's useful for users who are interested in QDoc errors
only. However, when called with this flag QDoc still creates output
directories, subdirectories (such as images, scripts, and style), and
auxiliary files (e.g. index, help project, manifest files).

This change adds guard clauses at the entry points of filesystem
operations to ensure complete suppression when the flag is set. By
employing early returns, QDoc avoids unnecessary work and does not
generate unexpected file system artifacts.

Certain warnings are generated during the generation phase. Running QDoc
with this flag skips that code path entirely. As a result, warnings
reported to stderr may be incomplete. This is expected to change at some
point in the future, as the generation step shouldn't be responsible for
issuing diagnostics.

Pick-to: 6.10
Fixes: QTBUG-140728
Change-Id: I6d022ff41fe05deb82c6d2d75a66b9ad5df97757
Reviewed-by: David Boddie <david.boddie@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>