qtattributionscanner: Perform package validation after filtering
When reading/parsing the files no checks should be performed and the
information just collected.
The filtering option is used to exclude certain packages from processing.
Yet the validation was still performed on all packages, which causes validation
errors even on packages that are to be excluded.
The validation step is being moved after the filtering of the packages so
that is only performed on the packages that are actually been processed.

Fixes: QTBUG-141619
Change-Id: Ibded97d85db6fb1fcb64a58c935e481c731b72f7
Reviewed-by: Kai Köhne <kai.koehne@qt.io>