lupdate: simplify C++ expression parsing
This patch changes the matchExpression method to keep parsing
tokens, until reaching the extra closing parentheses. This is
sufficient for lupdate, since the parser of lupdate is not
interested in knowing the exact structure of the expression. It
merely wants to know when the expression ends.

Also, rename matchExpression to skipExpression, which better
describes the new behavior.

Fixes: QTBUG-141631
Fixes: QTBUG-141632
Pick-to: 6.10
Change-Id: I254f6a3a1ceac5389e5564b8fb98e85099ba3cca
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>