QDoc: Allow custom enumerator prefixes in \qmlenum topic
When listing enumerators in a \qmlenum topic using \value commands,
QDoc always prefixed the enumerators with the name of the parent
QML type.

There are cases where this would result in incorrect documentation.
Notably, some types include enumerations from the Qt namespace,
where enumerators are prefixed with `Qt.`

Detect whether a \value argument includes a prefix, and if so, use it
as is.

Fixes: QTBUG-141401
Pick-to: 6.10
Change-Id: I26608ad074995a14c1c0e2646489b8e17c138a11
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: David Boddie <david.boddie@qt.io>